Double Checked on Stream-In Logic
More careful inspection on the S00 Axis interface, checked the delay
lucky enough to avoid one scanerio which didn't considered at the
beginning of design, which is the last batch of the input stream may
not be effectively be latched by the bram, for it ends at the buffer

but the cnt increment logic didn't check the input handshake signal,
so the cnt will increase anyways, then the latch batch will also be
latched correctly.

the S00 interface will delay the outside data and valid by one cycle
but the relative timing is correct. the ACCEP signal is full combinational
logic, so it will stop the incoming stream immediately, no data loss
hazard.

Refined the test bench as well, more similar to real situations.